Add files via upload
This project addresses the challenges of detecting evasive malware in PDF files by employing a stacking ensemble model that combines classifiers like Random Forest, XGBoost, LightGBM, and CatBoost, with a Gradient Boosting meta-learner. The methodology integrates PCA for dimensionality reduction, SMOTE for addressing class imbalances, and a preprocessing pipeline for handling diverse features. The system was evaluated on the Evasive-PDFMal2022 dataset, achieving a state-of-the-art accuracy of 99.89% and demonstrating high resilience against evasive malware attacks. This robust approach highlights the potential of advanced ensemble techniques in strengthening cybersecurity frameworks.